Enforce consistent docker image usage
This patch introduces .dockersum, a directory containing checksums for
the docker images used in build release and test.  This ensures
containers are consistent until intentionally updated.

The .drone.yml has been updated to take advantage of these consistent
images, ensuring CI/release is the same locally as it is remotely.